refactor: ドメインディレクトリを多段構造に再編（Rust 2018 スタイル）
- `policy.rs` を廃止し evaluate() を `pr_state.rs` に吸収
- `PrState` を `domain/pr_state.rs` に配置（発見性の向上）
- ディレクトリ構造で多段ドメインを表現（Rust 2018 edition スタイル）:
  - `domain/pr_state.rs`                  ← PrState / PrLifecycle / evaluate()
  - `domain/pr_state/blocked.rs`          ← BlockedState
  - `domain/pr_state/blocked/branch_sync.rs`
  - `domain/pr_state/blocked/ci.rs`
  - `domain/pr_state/blocked/review.rs`
  - `domain/pr_state/unblocked.rs`        ← UnblockedState / MergeReadiness / UnblockedRepository
- 旧 `branch_sync/`, `ci_checks/`, `review/`, `unblocked/` の各トップレベルモジュールを削除
- `domain.rs` を `pub mod error; pub mod pr_state;` のみに整理
