+++ src/contexts/evaluation/application/unblocked.rs
@@ -1,5 +1,7 @@
 use super::errors::{ErrorLogger, ErrorPresenter};
-use crate::contexts::evaluation::domain::unblocked::{MergeReadiness, UnblockedRepository};
+use crate::contexts::evaluation::domain::pr_state::unblocked::{
+    MergeReadiness, UnblockedRepository,
+};
 
 /// マージ可否状態を取得する。失敗時は `None` を返してエラー出力する。
 pub fn fetch(
