--- src/contexts/evaluation/domain/branch_sync.rs
@@ -1,5 +1,0 @@
-mod repository;
-mod status;
-
-pub use repository::BranchSyncRepository;
-pub use status::{BranchSync, BranchSyncState, BranchSyncStatus};
