--- src/contexts/evaluation/domain/ci_checks.rs
@@ -1,10 +1,0 @@
-mod aggregate;
-mod bucket;
-mod checks;
-mod repository;
-mod status;
-
-pub use bucket::CheckBucket;
-pub use checks::CiChecks;
-pub use repository::CiChecksRepository;
-pub use status::{CiState, CiStatus};
