--- src/contexts/evaluation/domain/ci_checks/aggregate.rs
@@ -1,19 +1,0 @@
-use super::{CheckBucket, CiStatus};
-
-/// 複数チェックを集約（優先度: `Fail` > `ActionRequired` > `Pass`）
-#[must_use]
-pub fn aggregate(buckets: &[CheckBucket]) -> CiStatus {
-    if buckets
-        .iter()
-        .any(|b| matches!(b, CheckBucket::Fail | CheckBucket::Cancel))
-    {
-        CiStatus::Fail
-    } else if buckets
-        .iter()
-        .any(|b| matches!(b, CheckBucket::ActionRequired))
-    {
-        CiStatus::ActionRequired
-    } else {
-        CiStatus::Pass
-    }
-}
