--- src/contexts/evaluation/domain/ci_checks/bucket.rs
@@ -1,9 +1,0 @@
-/// 個別チェックのバケット種別
-///
-/// `Fail`/`Cancel`/`ActionRequired` は `ci-fail`/`ci-action` 判定に使用する。
-pub enum CheckBucket {
-    Fail,
-    Cancel,
-    ActionRequired,
-    Other,
-}
