--- src/contexts/evaluation/domain/ci_checks/repository.rs
@@ -1,8 +1,0 @@
-use super::super::error::RepositoryError;
-use super::CiChecks;
-
-pub trait CiChecksRepository {
-    /// # Errors
-    /// Returns `RepositoryError` if the CI checks cannot be fetched.
-    fn fetch_checks(&self) -> Result<CiChecks, RepositoryError>;
-}
