--- src/contexts/evaluation/domain/ci_checks/status.rs
@@ -1,13 +1,0 @@
-/// チェック全体の集約状態（インフラから取得した生の値）
-pub enum CiStatus {
-    Fail,
-    ActionRequired,
-    Pass,
-}
-
-/// CI のブロッカー評価状態
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum CiState {
-    Fail,
-    ActionRequired,
-}
