--- src/contexts/evaluation/domain/review.rs
@@ -1,7 +1,0 @@
-mod model;
-mod repository;
-mod status;
-
-pub use model::Review;
-pub use repository::ReviewRepository;
-pub use status::{ReviewState, ReviewStatus};
