--- src/contexts/evaluation/domain/review/model.rs
@@ -1,23 +1,0 @@
-use super::{ReviewState, ReviewStatus};
-
-/// レビュー状態のドメインモデル
-pub struct Review {
-    status: ReviewStatus,
-}
-
-impl Review {
-    #[must_use]
-    pub fn new(status: ReviewStatus) -> Self {
-        Self { status }
-    }
-
-    #[must_use]
-    pub fn state(&self) -> Option<ReviewState> {
-        match self.status {
-            ReviewStatus::ChangesRequested => Some(ReviewState::ChangesRequested),
-            ReviewStatus::Approved | ReviewStatus::ReviewRequired | ReviewStatus::NoDecision => {
-                None
-            }
-        }
-    }
-}
