--- src/contexts/evaluation/domain/review/status.rs
@@ -1,13 +1,0 @@
-/// レビュー決定状態（インフラから取得した生の値）
-pub enum ReviewStatus {
-    ChangesRequested,
-    Approved,
-    ReviewRequired,
-    NoDecision,
-}
-
-/// レビューのブロッカー評価状態
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum ReviewState {
-    ChangesRequested,
-}
