--- src/contexts/evaluation/domain/unblocked/repository.rs
@@ -1,8 +1,0 @@
-use super::super::error::RepositoryError;
-use super::MergeReadiness;
-
-pub trait UnblockedRepository {
-    /// # Errors
-    /// Returns `RepositoryError` if the merge readiness cannot be fetched.
-    fn fetch_readiness(&self) -> Result<MergeReadiness, RepositoryError>;
-}
